% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2hex.R
\name{col2hex}
\alias{col2hex}
\title{col2hex}
\usage{
col2hex(color, alpha = 255, maxValue = 255)
}
\arguments{
\item{color}{- character}
}
\value{
none
}
\description{
Convert a color's character string to its hexadecimal code.
}
\examples{
col2hex('red', 204)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
