% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex2col.R
\name{fix_hex}
\alias{fix_hex}
\title{hex2col}
\usage{
fix_hex(hex_color)
}
\arguments{
\item{hex}{- hexadecimal color}
}
\value{
color - name of the color
}
\description{
get the approximate name of a hex color.
}
\examples{
hex2col('3F273C')
hex2col 3F273C # as executable

}
\references{
Silence Dogood, \StackOverflow{https://bit.ly/3gngca8}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
