% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrd.R
\name{getrd}
\alias{getrd}
\title{getrd - get a git repositories root directory}
\usage{
getrd(dpattern = ".git", fpattern = NULL, max_trys = 5)
}
\arguments{
\item{dpattern}{- directory pattern specifying root directory}

\item{fpattern}{- file pattern specifying project's root directory}

\item{max_tries}{- maximum depth of project tree}
}
\value{

}
\description{
get the root directory of a git project
}
\examples{
getrd()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{twesleyb10@gmail.com}
}
\keyword{}
