% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ggplotScaleFreePlot}
\alias{ggplotScaleFreePlot}
\title{ggplotScaleFreePlot}
\usage{
ggplotScaleFreePlot(connectivity, nBreaks = 10, truncated = FALSE,
  removeFirst = FALSE, main = "", ...)
}
\arguments{
\item{connectivity}{- igraph connectivity of a graph}
}
\value{
None
}
\description{
evaluate the scale free property of a graph
}
\examples{

}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{connectivity}
\keyword{igraph}
\keyword{scale-free-fit}
