% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsavePDF.R
\name{ggsavePDF}
\alias{ggsavePDF}
\title{ggsavePDF}
\usage{
ggsavePDF(plots, file)
}
\value{
none
}
\description{
Function for saving multiple ggplots to single pdf.
}
\examples{
ggsavePDF(plots, file)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
