% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsaveTable.R
\name{ggsaveTable}
\alias{ggsaveTable}
\title{ggsaveTable}
\usage{
ggsaveTable(mytable, myfile)
}
\arguments{
\item{mytable}{a grob object}
}
\description{
save a table cropped to appropriate height and width dimensions
}
\references{
\url{https://stackoverflow.com/questions/13867325/get-width-of-plot-area-in-ggplot2}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{grob}
\keyword{height}
\keyword{size}
\keyword{width}
