% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gitadd}
\alias{gitadd}
\title{git add}
\usage{
gitadd(dir = getwd())
}
\arguments{
\item{dir}{character specifying the directory of a git repository.}
}
\value{
None
}
\description{
Add file contents to the git index.
}
\examples{
gitadd()

}
\references{
\url{https://git-scm.com/docs/git-add}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{add}
\keyword{bash}
\keyword{cmd}
\keyword{git}
