% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gitcommit}
\alias{gitcommit}
\title{git commit}
\usage{
gitcommit(msg = "commit from Rstudio", dir = getwd())
}
\arguments{
\item{msg}{character describing changes made to the repository.}

\item{dir}{character specifying the directory of a git repository.}
}
\value{
None
}
\description{
Record changes to the git repsoitory. Note: the user should have previously
  called git2r::config(user.name,user.email) for git to recognize R.
}
\examples{
gitcommit()

}
\references{
\url{https://git-scm.com/docs/git-commit}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{bash}
\keyword{cmd}
\keyword{commit}
\keyword{git}
