% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gitconfig}
\alias{gitconfig}
\title{git config}
\usage{
gitconfig(user_name, user_email)
}
\arguments{
\item{user_name}{the user's GitHub username.}

\item{user_email}{the user's GitHub email.}
}
\value{
None
}
\description{
Configure git..
}
\examples{
gitconfig("twesleyb", "tyler.w.bradshaw@duke.edu")
}
\references{
\url{https://git-scm.com/docs/git-config}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{bash}
\keyword{cmd}
\keyword{config}
\keyword{git}
