% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gitpull}
\alias{gitpull}
\title{git pull}
\usage{
gitpull(dir = getwd())
}
\arguments{
\item{dir}{character specifying the directory of a git repository.}
}
\value{
None
}
\description{
Incorporate changes from remote git repository to current, local branch.
}
\examples{
gitpull()
}
\references{
\url{https://git-scm.com/docs/git-pull}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{bash}
\keyword{cmd}
\keyword{git}
\keyword{status}
