% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gitpush}
\alias{gitpush}
\title{git push}
\usage{
gitpush(dir = getwd())
}
\arguments{
\item{dir}{character specifying the directory of a git repository.}
}
\value{
None
}
\description{
Update remote git repository with local changes.
}
\examples{
gitpush()

}
\references{
\url{https://git-scm.com/docs/git-push}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{bash}
\keyword{cmd}
\keyword{git}
\keyword{push}
