% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gitstatus}
\alias{gitstatus}
\title{git status}
\usage{
gitstatus(dir = getwd())
}
\arguments{
\item{dir}{character specifying the directory of a git repository.}
}
\value{
None
}
\description{
Check git status of a repository.
}
\examples{
gitstatus()

}
\references{
\url{https://git-scm.com/docs/git-status}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{bash}
\keyword{cmd}
\keyword{git}
\keyword{status}
