% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{notin}
\usage{
... \%notin\% NA
}
\value{
none
}
\description{
%notin% - the opposite of %in%
}
\examples{
x \%notin\% y

}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
