% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{grobsize}
\alias{grobsize}
\title{grobsize}
\usage{
grobsize(x)
}
\arguments{
\item{x}{(grob) a grob object}
}
\value{
None
}
\description{
get the actual height and width of a grob
}
\examples{
ggsize(grob)

}
\references{
\url{https://stackoverflow.com/questions/13867325/get-width-of-plot-area-in-ggplot2}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{grob}
\keyword{height}
\keyword{size}
\keyword{width}
