% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex2name.R
\name{hex2name}
\alias{hex2name}
\title{hex2name}
\usage{
hex2name(hex)
}
\arguments{
\item{hex}{- hexadecimal color}
}
\value{
color - name of the color
}
\description{
get the approximate name of a hex color.
}
\references{
[Silence Dogood](https://bit.ly/3gngca8)
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
