% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex2str.R
\name{hex2str}
\alias{hex2str}
\title{hex2str}
\usage{
hex2str(hex)
}
\arguments{
\item{hex}{- hexadecimal color}
}
\value{
color - name of the color
}
\description{
get the approximate name of a hex color.
}
\references{
[Silence Dogood](https://bit.ly/3gngca8)
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
