% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_str.R
\name{knit_str}
\alias{knit_str}
\title{knit_str}
\usage{
knit_str(robj, ...)
}
\value{
knits the output of str
}
\description{
knitr::kable(str(robj))
}
\references{
[StackOverflow](https://stackoverflow.com/questions/44200394)
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
