% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex2col.R
\name{rgb2hex}
\alias{rgb2hex}
\title{hex2col}
\usage{
rgb2hex(r, g, b)
}
\arguments{
\item{hex}{- hexadecimal color}
}
\value{
color - name of the color
}
\description{
get the approximate name of a hex color.
}
\examples{
hex2col('3F273C')
hex2col 3F273C # as executable

}
\references{
[Silence Dogood](https://bit.ly/3gngca8)
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
