% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2name.R
\name{rgb2name}
\alias{rgb2name}
\title{rgb2name}
\usage{
rgb2name(r, g, b, show = FALSE)
}
\arguments{
\item{r}{- red}

\item{g}{- green}

\item{b}{- blue}
}
\value{
color - name of the color
}
\description{
get the approximate name of a rgb color.
}
\references{
[Silence Dogood](https://bit.ly/3gngca8)
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
