% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2str.R
\name{rgb2str}
\alias{rgb2str}
\title{rgb2str}
\usage{
rgb2str(r, g, b, show = FALSE)
}
\arguments{
\item{r}{- red}

\item{g}{- green}

\item{b}{- blue}
}
\value{
color - name of the color
}
\description{
get the approximate name of any rgb color.
}
\references{
[Silence Dogood](https://bit.ly/3gngca8)
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
