% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rip}
\alias{rip}
\title{rip}
\usage{
rip(package, method = "utils", ...)
}
\arguments{
\item{package}{}
}
\value{
none
}
\description{
Utility to help install R packages.
}
\examples{
rip("package", method = "BiocManager")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{install}
\keyword{install.packages}
\keyword{package}
