% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{silently}
\alias{silently}
\title{silently}
\usage{
silently(func, ...)
}
\arguments{
\item{func}{(function) symmetric adjacency matrix representing the network graph.}

\item{...}{(string) additional arguments passed to func().}
}
\value{
None
}
\description{
suppress any unwanted output from a function with sink().
}
\examples{
silently(wgcna::bicor, exprDat)

}
\references{
\url{}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{output}
\keyword{quiet}
\keyword{silent}
\keyword{supress}
