% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{touch}
\alias{touch}
\title{Creating a file}
\usage{
touch(file = "test.txt", dir = getwd())
}
\arguments{
\item{file}{character specifying file that will be created.}

\item{dir}{character specifying the working direcotry.}
}
\value{
None
}
\description{
Creates a file in the current working directory. Inspired by the linux command `touch`.
}
\examples{
touch("test.txt")

}
\references{
\url{https://gist.github.com/JoshCheek/1224782}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{create}
\keyword{directory}
\keyword{file}
\keyword{linux}
\keyword{touch}
