% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{uniprot_mapping}
\alias{uniprot_mapping}
\title{uniprot_mapping}
\usage{
uniprot_mapping(ids, from = "ACC+ID", to = "GENENAME")
}
\arguments{
\item{ids}{- gene identifiers}

\item{from}{- input format}

\item{to}{- output format}
}
\value{
gene identifiers returned from Uniprot mapping service.
}
\description{
Query the Uniprot database mapping service in order to map gene ids.
}
\examples{
uniprot_mapping("ROGDI_HUMAN","ACC+ID","P_ENTREZGENEID")
}
\references{
\url{https://www.uniprot.org/help/api_idmapping}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{color}
\keyword{convert}
\keyword{hex}
\keyword{rgb}
