% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_size.R
\name{workspace.size}
\alias{workspace.size}
\title{workspace_size}
\usage{
workspace.size()
}
\value{
the size of the current environment in bytes
}
\description{
workspace_size
}
\references{
\url{https://stackoverflow.com/questions/22023404}
}
\author{
sgibb
}
