% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{write.pajek}
\alias{write.pajek}
\title{write.pajek}
\usage{
write.pajek(adjm, file = "network.net", col.names = FALSE, ...)
}
\arguments{
\item{adjm}{(matrix) symmetric adjacency matrix representing the network graph.}

\item{file}{(string) name of output file (e.g. 'network.net')}
}
\value{
None
}
\description{
Write network adjacency network to file in Pajek (*.net) format.
}
\examples{
write.pajek(adjm, "network.net")

}
\references{
\url{https://gephi.org/users/supported-graph-formats/pajek-net-format/}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{graph}
\keyword{network}
\keyword{pajek}
\keyword{write}
