% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_excel.R
\name{write_excel}
\alias{write_excel}
\title{Create an excel workbook.}
\usage{
write_excel(data, file, rowNames = FALSE, colNames = TRUE, ...)
}
\arguments{
\item{data}{- named list of data frames (or equivalent) to be written to
an excel document.}

\item{file}{- output file name.}

\item{rowNames}{- should rownames of data be kept?}

\item{colNames}{- should colnames of data be kept?}

\item{'...'}{- additional arguments passed to openxlsx::writeData()}
}
\value{
none
}
\description{
Create an excel workbook.
}
\examples{
write.excel(data, file = "foo.xlsx")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{document}
\keyword{excel}
\keyword{write}
\keyword{write.excel}
\keyword{xls}
\keyword{xlsx}
