% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{col2hex}
\alias{col2hex}
\title{col2hex}
\usage{
col2hex(color, maxValue = 255)
}
\arguments{
\item{color}{- the color to be converted to hex.}

\item{maxValue}{- number giving the maximum of the color values range, see above.}
}
\value{
None
}
\description{
convert a color to hexadecimal code
}
\examples{
col2hex("red")

}
\references{

}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
