% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormind.R
\name{colormind}
\alias{colormind}
\title{colormind}
\usage{
colormind(c1 = NA, c2 = NA, c3 = NA, c4 = NA, c5 = NA)
}
\arguments{
\item{c1}{- a color passed as a rgb vector, e.g. c(85,90,84).}

\item{c2}{- a color passed as a rgb vector, e.g. c(85,90,84).}

\item{c3}{- a color passed as a rgb vector, e.g. c(85,90,84).}

\item{c4}{- a color passed as a rgb vector, e.g. c(85,90,84).}

\item{c5}{- a color passed as a rgb vector, e.g. c(85,90,84).}
}
\description{
Query the colormind API for a random color palette. Seed colors can be 
specified by passing up to five rgb colors to the function.
}
\references{
\href{http://colormind.io/}{colormind}
}
