% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrd.R
\name{getrd}
\alias{getrd}
\title{getrd - get the root directory}
\usage{
getrd(dpattern = ".git", fpattern = NULL, max_trys = 5)
}
\arguments{
\item{none}{}
}
\value{

}
\description{
get the root directory of a git project.
}
\examples{
getrd()

}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
