% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ggplotScaleFreePlot}
\alias{ggplotScaleFreePlot}
\title{ggplotScaleFreePlot}
\usage{
ggplotScaleFreePlot(connectivity, nBreaks = 10, truncated = FALSE,
  removeFirst = FALSE, main = "", ...)
}
\arguments{
\item{x}{(grob) a grob object}
}
\value{
None
}
\description{
get the actual height and width of a grob
}
\examples{
ggsize(grob)

}
\references{
\url{https://stackoverflow.com/questions/13867325/get-width-of-plot-area-in-ggplot2}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{grob}
\keyword{height}
\keyword{size}
\keyword{width}
