% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2hex.R
\name{rgb2hex}
\alias{rgb2hex}
\title{rgb2hex}
\usage{
rgb2hex(r, g, b)
}
\arguments{
\item{r, }{g, b - numeric red, green, and blue components of a color}
}
\value{
hexadecimal code
}
\description{
Convert a color's rgb representation to its hexadecimal code.
}
\references{
https://gist.github.com/mbannert/e9fcfa86de3b06068c83
}
\author{
Hadley Wickam
}
\keyword{none}
