% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rip}
\alias{rip}
\title{rip}
\usage{
rip(package, method = "utils", ...)
}
\arguments{
\item{package}{- package to be installed.}

\item{method}{- method to be used.}
}
\value{
none
}
\description{
Utiliity function to help install R packages.
}
\examples{
rip("https://cran.r-project.org/src/contrib/ncdf4_1.17.tar.gz",method="source")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{install}
\keyword{install.package}
\keyword{package}
