% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.size.R
\name{workspace.size}
\alias{workspace.size}
\title{Create an excel workbook.}
\usage{
workspace.size()
}
\value{
the size of the current environment in bytes
}
\description{
Create an excel workbook.
}
\references{
\url{https://stackoverflow.com/questions/22023404}
}
\author{
sgibb
}
