% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{write.excel}
\alias{write.excel}
\title{write.excel}
\usage{
write.excel(data, file, rowNames = FALSE, colNames = TRUE)
}
\arguments{
\item{data}{- the named list of dataframes.}
}
\value{
None
}
\description{
Writes a list of data frames to excel workbook with the openxlsx library.
}
\examples{
write.excel(data, file = "foobar.xlsx")

}
\references{

}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
