% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{write.excel}
\alias{write.excel}
\title{Write data to excel workbook.}
\usage{
write.excel(data, file, rowNames = FALSE, colNames = TRUE)
}
\arguments{
\item{data}{- list of data frames to be written to an excel document.}
}
\value{
none
}
\description{
Write data to excel workbook.
}
\examples{
write.excel(data,file="foo.xlsx")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{documnent}
\keyword{excel}
\keyword{write}
\keyword{write.excel}
\keyword{xls}
\keyword{xlsx}
